290.1. Rendering Authentication Status(Section 21: Handling Authentication with React)
In the last video we played around with the API library a little bit more inside the browser console.

We saw that we can call that get off instance method.

And anytime that we call that it's going to give us a object back that has a ton of different functions

assigned to it that we can use to essentially manipulate the user's authentication status.

So now in this video we're going to continue developing our Google Earth class right here.

Let me set up a couple of quick goals for what we're going to try to do.

All right.

So inside of my component or the Google off component I want to try to get a reference to that off object

after it is initialized.

Now when I say object right here I'm talking about how just in the last video we were able to say API

to get off instance when we call that this is the author object I'm talking about.

So this is the thing that has all those different methods assigned to it that we can use to manipulate

the user's authentication status or even figure out if they are currently signed in.

So we're going to get a reference to that object after we initialize the library.

Then after that we're going to attempt to figure out whether or not the user is currently signed in.

Remember in the last video we went through the authentication flow.

So right now in theory we are currently a lot with Google inside of our application.

So I want to figure out whether or not the user using our application is signed in.

And then once we figure that out I want to somehow print out that authentication status on the screen.

It's just a very quick little goal here to make sure that we understand the basics of Google both in

how to interact with this library.

Remember right now we're not going to worry about redux at all.

I just want to get this component up and running by itself.

OK.

So back inside of my code editor I'm going to find my google off class.

So remember when we loaded up that additional library right here we had to pass in a callback function

that was invoked after this additional module inside the library was success successfully loaded up

in a very similar fashion.

When we call client Dom init right here that executes an asynchronous operation or an asynchronous network

request over to Google's API server in order to initialize our client.

So we want to get some type of callback function or some type of notice for when that initialization

set up is all done.

So to do so we're going to chain on a then statement right here.

All right let me tell you why we are doing that.

When we called Load up here load only allows us to get a signal or a notification of when the loading

process is complete bypassing in a callback function.

But when we call init we do not have to use a callback function when we called in it it returns a promise

a promise is an object that will give us a little tap on the shoulder so to speak after this client

library has been successfully initialized.

So by chaining on this.

Then steepen right here and passing in an Aero function.

This aero function is going to be automatically invoked after a library has successfully initialized

itself.

So inside of here we can write some amount of code that will be only executed once our entire API library

is ready to go.

So it's at this point right here that we can start to figure out whether or not the user is currently

signed in and then attempt to print out that status on the screen.

So inside if you're going to first begin by getting a reference to that off object and saving a reference

to it on my component class.

So I'm going to say this dot off equals window dot G API.

Off to get off instance.

So now after running this code right here inside of any other function inside my class I can reference

this dot off and that will give me a reference to that Auth. instance that I can use to sign the user

in or sign them out or get the users current on the nation status.

Because remember our goal right now is to somehow figure out whether or not the user is signed in and

then print out that authentication status on the screen.

So keep in mind that when this component or this method right here gets called our component has already

been rendered to the screen.

So now if we want to essentially update the text or what content this component shows we need to somehow

get the component to the render.

So we're going to be making use of component level state here.