290.2. Rendering Authentication Status(Section 21: Handling Authentication with React)
After we initialize our library and get access to this object we're going to update our component level

state with whether or not the users currently signed in when we update that state.

It will cause the component to automatically render and we can then print up the authentication status

on the screen so inside of here after getting a reference to this start off I'm going to call this Scott

set state and I going to pass an object here and say is signed in.

And that's going to be the result of calling this off that is signed in.

Get like so.

OK so now after loading a library after initializing it we finally get our reference to the author library

itself or the author object.

And then we're going to update our component level state with this new property called is signed in.

And we're just going to take that directly off with that object now is signed in right here if you want

to you can flip back over to the Google API documentation and see that this is a property on that object.

And if we want to get whether or not the user is currently signed in we just referenced that get function

or we invoke it like so.

OK so now that we are making use of component level state we need to make sure that we initialize the

state object when our component class is created.

So at the top of this class I'm going to say state is is signed in and I'm going to default that to

have a value of NULL.

The reason I'm putting no right here is that we do not know if the user is signed in or not signed in

when our application first loads.

So if there is a value of no four is signed in that means that we do not know if the user is signed

in or not signed in.

And so we we probably should not print anything on the screen at that point in time.

OK.

So now inside my render method I could put some logic directly inside if you're to look at the value

of this statement is signed in and then print out either yes no they are signed in.

But instead I'm going to make a little helper method just to help us kind of print out the appropriate

text sanguinea make a helper method called render off button.

Right now we're just going to be returning some div with text in it.

But as you might guess in just a moment we're going to replace this with some actual button logic or

some button to show up on the screen.

OK so inside if you're all say if this statement is signed in equals no then I'm going to return a div

that says I don't know if we are signed

and then else if this state DOT this statement is signed in we could say equals equals true.

But we don't actually have to say equals equals true.

We can just reference the property by itself and would be the same thing.

Now if we are signed in then I'm going to return a div that says I am signed in.

And then finally we'll handled the last case the case in which I am not signed in and is signed in is

equal to a value of false.

And so in this case I will return a div that says I am not signed in like so.

And then finally done inside my render method I'm going to reference that helper method we just put

together.

So I will call this render off

OK.

I'm going to save this and we'll flip back with the browser and test this out.

So I'm going to flip back over and I see at the top right hand side it says I am signed in.

So that means we are currently signed in.

Now just to make this a little bit more interesting let's try actually signing out right now we don't

have a button to sign our user out but we can very easily sign out by using our counsel and directly

manipulating the API library so I can say gee API.

Off to get off instance your number that's going to get us access to the thought ject.

And then one of the methods on that object is sign out right there.

So I'm going to say gee AAPI off to get instant start.

Sign out.

And then once I called out I'm no longer considered to be signed in via Google Earth so I can now refresh

the page.

And when I do so I should see I am not signed in print up inside my head or.

So refresh of the page.

You see that it said there very briefly it said I don't know if I'm signed in.

That was while the JPI Libre was loading up and trying to decide whether or not our user was signed

in and the ones that determined that the user was not signed in we called Set States and we now see

I am not signed in.

So now if we want to we can sign in one more time.

By doing GPI off to get instance sign in that will kick us through the floor again.

So I'm going to select an account right here.

And then we get kicked active back over to our page.

Now we don't have any logic inside of our application right now to get a callback or some type of notification

when the user does sign in.

So if you want to see that header text update we have to actually refresh the page.

So I'll do so right now and you'll see that it now goes back to saying I am signed in.

Now the really nice thing about this Google API library is that it's going to hold your authentication

status between refreshes of the page.

So for example if I refresh right now you'll see that I am still signed in and I can refresh as much

as I want and I'm still going to be signed in until I actually sign out.

OK.

So that's pretty much the basics of this library.

In a nutshell.

Now it would be really nice if we made sure that whenever the user actually signs in we dynamically

like on the fly changed the text right here.

So let's try to take a shot at that in the next section.

So quick break and I'll see you in just a minute.