291.2. Updating Auth State(Section 21: Handling Authentication with React)
Now one little optimization we could do here.

You'll notice how inside of the dot then statement we have essentially the same exact line of code that

we have down inside of on change.

So if we wanted to we could actually remove this right here and instead call this dot on off change

like so that's going to make sure that we get the current authentication status when we first initialize

our library and update the authentication state to be whatever its values should be.

So if you want to do that refactor it's fine.

So let's take care of that in the next video.