291.3. Updating Auth State(Section 21: Handling Authentication with React)
But I would recommend you kind of leave it as is for right now because we're going to eventually come

back to this class into a big refactor to it or not a big factor but a little refactor once we add redux

into our project.

So I would recommend you have the tools set state calls for right now even though it's not super ideal.

OK so this looks great.

Now the last thing we do is make sure that our render off button right here doesn't return as simple

div but instead returns an actual off button that says the user to the user either hey click this to

sign in or click this to sign out.

So let's take care of that in the next video.